Severity monoid
This patch introduces a commutative monoid instance for Severity which
should be used for any severity calucations.  Unfortuntately it is not
possible to remove `Ord` instance, which can be confused.  This patch
also fixes `testSeverity` computation.  It should compare the message
severity with minimum severity found in any configuration settings.
This allows to turn on messages from a tracer  for better dubugging
information, rather than turn them off.  That seems like a more
sensible default.