severities:  restore semantics of #593 and update tests accordingly
Previously, before #593, to determine whether a message passes
the severity threshold, the 'testSeverity' predicate was
computing a maximum of global and per-trace severities.

Therefore, originally, there was no way to enable a particular tracer,
although it was possible to mute one.

So, #593 fixed that, by making `testSeverity` compare messages
against the _minimum_ of global and per-trace severities.

That PR was temporarily reverted due to failing tests and
my confusion.

Thanks to @coot for digging into the details here!