Add ToObject requirement and tests
This changes the output of `ScText`-configured loggers to use the
'ToObject' typeclass to get the text representation. In particular, we
now use the 'textTransformer' function to get the output, which by
default comes from the ToJSON instance, but can be customised as one may
wish (in particular, for example, to be the output of a pretty-printer).